Misc Changes Removed old lab preamble comments from usart code Made `GcodeArg.value` a 32 bit int (I don't know why I made it 16 bit before) Added a more generic `assert_return` macro in the hopes of eventually replacing the `gcode_assert` macro with something more reusable and generic.

+++ src/gcodes.h
@@ -10,7 +10,7 @@
 typedef struct GcodeArg {
 	char id;
 	// TODO: Make it so value can be either an int or a float
-	uint16_t value;
+	int32_t value;
 } GcodeArg;
 
 typedef struct Gcode {
